+++ src/it/java/org/greencodeinitiative/creedengo/java/integration/tests/GCIRulesIT.java
@@ -549,10 +549,10 @@ class GCIRulesIT extends GCIRulesBase {
     }
 
     @Test
-    void testGCI96_1() {
-        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidRuntimeExceptions.java";
+    void testGCI96_nominalRuntimeException() {
+        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidCatchingRuntimeExceptions.java";
         String ruleId = "creedengo-java:GCI96";
-        String ruleMsg = "Avoid Runtime exceptions : RuntimeException";
+        String ruleMsg = "Avoid catching Runtime exceptions : RuntimeException";
         int[] startLines = new int[]{36};
         int[] endLines = new int[]{37};
 
@@ -560,10 +560,10 @@ class GCIRulesIT extends GCIRulesBase {
     }
 
     @Test
-    void testGCI96_2() {
-        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidRuntimeExceptions.java";
+    void testGCI96_IndexOutOfBoundsException() {
+        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidCatchingRuntimeExceptions.java";
         String ruleId = "creedengo-java:GCI96";
-        String ruleMsg = "Avoid Runtime exceptions : IndexOutOfBoundsException";
+        String ruleMsg = "Avoid catching Runtime exceptions : IndexOutOfBoundsException";
         int[] startLines = new int[]{44};
         int[] endLines = new int[]{45};
 
@@ -571,10 +571,10 @@ class GCIRulesIT extends GCIRulesBase {
     }
 
     @Test
-    void testGCI96_3() {
-        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidRuntimeExceptions.java";
+    void testGCI96_NullPointerException() {
+        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidCatchingRuntimeExceptions.java";
         String ruleId = "creedengo-java:GCI96";
-        String ruleMsg = "Avoid Runtime exceptions : NullPointerException";
+        String ruleMsg = "Avoid catching Runtime exceptions : NullPointerException";
         int[] startLines = new int[]{52};
         int[] endLines = new int[]{53};
 
@@ -582,10 +582,10 @@ class GCIRulesIT extends GCIRulesBase {
     }
 
     @Test
-    void testGCI96_4() {
-        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidRuntimeExceptions.java";
+    void testGCI96_ArithmeticException() {
+        String filePath = "src/main/java/org/greencodeinitiative/creedengo/java/checks/AvoidCatchingRuntimeExceptions.java";
         String ruleId = "creedengo-java:GCI96";
-        String ruleMsg = "Avoid Runtime exceptions : ArithmeticException";
+        String ruleMsg = "Avoid catching Runtime exceptions : ArithmeticException";
         int[] startLines = new int[]{59};
         int[] endLines = new int[]{60};